Implement setting X backgrounds without feh.
Implement a cache for finalized images. This is significantly faster than feh even without counting the overhead
of writing to disk before reading from it. Compared to a week ago a 2s
feedback loop in interactive mode is now 300-500ms.

Setting wallpapers in place of feh in random mode has gone from 750ms-1s
down to 270-400ms. The rest of the process is only 200-250ms.

This removes the restore file (~/.fehbg) but that's not too hard to
reimplement.

Switch from double boxing to Cow, making the code nicer to work with.