Async conversion of wallpaper setting code, saves 40-50ms from disk.
Code is a bit messier but not too much more after removing some old
code.

Close the shuffler in another thread once it's no longer used in
ramdom. Saves 60-70ms in random.

Do not read the resolution of images that are already fully cached.
Saves 60-100ms in random and 8+ seconds in sync mode.

All told with 1600 wallpapers over NFS, random is ~200ms faster and sync
is only 150-250ms instead of 8.5s+. Preview and interactive weren't
helped much by this.